requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.5
openpyxl==3.1.5
gspread==6.1.4
